Revert "use uint64 for values"
This reverts commit 798767f2f389f647c8735e88d8851d79ea21ad57.

+++ wire/msgtxsyscoin.go
@@ -10,7 +10,7 @@ import (
 
 type AssetOutType struct {
 	N uint32
-	ValueSat uint64
+	ValueSat int64
 }
 type AssetAllocationType struct {
 	VoutAssets map[uint32][]AssetOutType
@@ -23,9 +23,9 @@ type AssetType struct {
 	Symbol string
 	PubData []byte
 	PrevPubData []byte
-	Balance uint64
-	TotalSupply uint64
-	MaxSupply uint64
+	Balance int64
+	TotalSupply int64
+	MaxSupply int64
 	Precision uint8
 	UpdateFlags uint8
 	PrevUpdateFlags uint8
@@ -186,19 +186,18 @@ func (a *AssetType) Deserialize(r io.Reader) error {
 	if err != nil {
 		return err
 	}
-	a.Balance = DecompressAmount(valueSat)
+	a.Balance = int64(DecompressAmount(valueSat))
 
 	valueSat, err = ReadUint(r)
 	if err != nil {
 		return err
 	}
-	a.TotalSupply = DecompressAmount(valueSat)
-	
+	a.TotalSupply = int64(DecompressAmount(valueSat))
 	valueSat, err = ReadUint(r)
 	if err != nil {
 		return err
 	}
-	a.MaxSupply = DecompressAmount(valueSat)
+	a.MaxSupply = int64(DecompressAmount(valueSat))
 
 	return nil
 }
@@ -241,15 +240,15 @@ func (a *AssetType) Serialize(w io.Writer) error {
 	if err != nil {
 		return err
 	}
-	err = PutUint(w, CompressAmount(a.Balance))
+	err = PutUint(w, CompressAmount(uint64(a.Balance)))
 	if err != nil {
 		return err
 	}
-	err = PutUint(w, CompressAmount(a.TotalSupply))
+	err = PutUint(w, CompressAmount(uint64(a.TotalSupply)))
 	if err != nil {
 		return err
 	}
-	err = PutUint(w, CompressAmount(a.MaxSupply))
+	err = PutUint(w, CompressAmount(uint64(a.MaxSupply)))
 	if err != nil {
 		return err
 	}
@@ -318,7 +317,7 @@ func (a *AssetOutType) Serialize(w io.Writer) error {
 	if err != nil {
 		return err
 	}
-	err = PutUint(w, CompressAmount(a.ValueSat))
+	err = PutUint(w, CompressAmount(uint64(a.ValueSat)))
 	if err != nil {
 		return err
 	}
@@ -336,7 +335,7 @@ func (a *AssetOutType) Deserialize(r io.Reader) error {
 	if err != nil {
 		return err
 	}
-	a.ValueSat = DecompressAmount(valueSat)
+	a.ValueSat = int64(DecompressAmount(valueSat))
 	return nil
 }
